Modified the estimation phase.
For running the estimation phase (how many iterations are performed) the `get_callback()` function was not called with an activated setup context.
this is different from the actual measuring loop, where `get_callback()` was called within the setup context.
This might lead to different behaviours.
This commit makes sure that the estimation takes place under an activated setup context.